HeartDiseaseAnalysis/WebApp/requirements.txt
cachetools==4.1.1
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
Flask==1.1.2
google-api-core==1.22.4
google-api-python-client==1.12.8
google-auth==1.22.1
google-auth-httplib2==0.0.4
google-cloud-core==1.4.3
google-cloud-storage==1.31.1
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.3
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.4
requests==2.24.0
rsa==4.6
six==1.15.0
uritemplate==3.0.1
urllib3==1.25.11
Werkzeug==1.0.1
